+++ embulk-input-mysql/src/test/java/org/embulk/input/mysql/IncrementalTest.java
@@ -3,7 +3,6 @@ package org.embulk.input.mysql;
 import org.embulk.config.ConfigDiff;
 import org.embulk.config.ConfigSource;
 import org.embulk.formatter.csv.CsvFormatterPlugin;
-import org.embulk.input.MySQLInputPlugin;
 import org.embulk.input.file.LocalFileInputPlugin;
 import org.embulk.output.file.LocalFileOutputPlugin;
 import org.embulk.parser.csv.CsvParserPlugin;
