+++ embulk-input-mysql/src/test/java/org/embulk/input/mysql/OptionTest.java
@@ -2,8 +2,6 @@ package org.embulk.input.mysql;
 
 import org.embulk.config.ConfigSource;
 import org.embulk.formatter.csv.CsvFormatterPlugin;
-import org.embulk.input.MySQLInputPlugin;
-import org.embulk.input.MySQLInputPlugin;
 import org.embulk.input.file.LocalFileInputPlugin;
 import org.embulk.output.file.LocalFileOutputPlugin;
 import org.embulk.parser.csv.CsvParserPlugin;
