+++ embulk-input-postgresql/src/test/java/org/embulk/input/postgresql/ArrayTest.java
@@ -2,7 +2,6 @@ package org.embulk.input.postgresql;
 
 import org.embulk.config.ConfigSource;
 import org.embulk.formatter.csv.CsvFormatterPlugin;
-import org.embulk.input.PostgreSQLInputPlugin;
 import org.embulk.input.file.LocalFileInputPlugin;
 import org.embulk.output.file.LocalFileOutputPlugin;
 import org.embulk.parser.csv.CsvParserPlugin;
