+++ embulk-input-sqlserver/src/main/java/org/embulk/input/sqlserver/SQLServerInputPlugin.java
@@ -1,4 +1,4 @@
-package org.embulk.input;
+package org.embulk.input.sqlserver;
 
 import java.sql.Connection;
 import java.sql.Driver;
