+++ embulk-input-sqlserver/src/test/java/org/embulk/input/sqlserver/BasicTest.java
@@ -10,7 +10,6 @@ import java.nio.file.Path;
 import org.embulk.config.ConfigDiff;
 import org.embulk.config.ConfigSource;
 import org.embulk.formatter.csv.CsvFormatterPlugin;
-import org.embulk.input.SQLServerInputPlugin;
 import org.embulk.input.file.LocalFileInputPlugin;
 import org.embulk.output.file.LocalFileOutputPlugin;
 import org.embulk.parser.csv.CsvParserPlugin;
